divded reducer and fixed bugs with devtools-redux, response
